pyproject.toml
[tool.mypy]
ignore_missing_imports = true
disable_error_code = ["return-value", "operator"]

[tool.pytest.ini_options]
testpaths = ["scripts"]
pythonpath = ["."]

[project]
name = "ez-agentic-study-buddy"
version = "0.1.1"
description = "A single-agent AI tutor with autonomous learning path planning and adaptive teaching."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.0",
    "langchain-groq>=0.1.0",
    "langchain-openai>=1.1.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.22",
]
